Move FixLocation extension into a separate file.
