+++ MapboxCoreNavigation/MBNavigator.swift
@@ -1,18 +1,5 @@
 import Foundation
 import MapboxNavigationNative
-import CoreLocation
-
-extension FixLocation {
-    convenience init(_ location: CLLocation) {
-        self.init(coordinate: location.coordinate,
-                  time: location.timestamp,
-                  speed: location.speed >= 0 ? location.speed as NSNumber : nil,
-                  bearing: location.course >= 0 ? location.course as NSNumber : nil,
-                  altitude: location.altitude as NSNumber,
-                  accuracyHorizontal: location.horizontalAccuracy >= 0 ? location.horizontalAccuracy as NSNumber : nil,
-                  provider: nil)
-    }
-}
 
 /// :nodoc:
 extension RouteState: CustomStringConvertible {
